requirements.txt
asyncqtpy>=0.10.0
BlurWindow>=1.2.1
darkdetect>=0.8.0
EMD-signal>=1.6.0
Jinja2>=3.1.4
graphviz>=0.20.3
imblearn~=0.0
ipython~=8.26.0
lmfit==1.3.1
matplotlib>=3.9.0
mpl-axes-aligner>=1.3
numpy==1.26.4
numba>=0.60.0
optuna>=3.6.1
optuna-integration>=3.6.0
openpyxl>=3.1.5
pandas>=2.2.2
psutil>=5.9.8
pybaselines>=1.1.0
pyqtgraph>=0.13.7
pyperclip>=1.8.2
PyQt5>=5.15.10
PyQt5-Frameless-Window>=0.3.9
PyQtWebEngine>=5.15.6
PyQtWebEngine-Qt5>=5.15.2
PyYAML>=6.0.1
QtPy>=2.4.1
scipy>=1.13.1
shap>=0.45.1
seaborn>=0.13.2
setuptools>=70.0.0
xgboost>=2.1.0
scikit-learn>=1.5.1
hyperopt>=0.2.7
yappi>=1.3.6
rampy>=0.4.9
BaselineRemoval>=0.1.4
numdifftools>=0.9.41
emcee>=3.1.4
dash>=2.9.2
traitlets~=5.11.2
